{{cookiecutter.repository_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.package_name}}"
version = "0.0.0"
description = "Placeholder description"
authors = ["{{cookiecutter.author}}"]
readme = "README.rst"
repository = "{{cookiecutter.repository}}"
# some classifiers are added automatically, see https://python-poetry.org/docs/pyproject/#classifiers
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Other Environment",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "{{cookiecutter.package_name}}" },
]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.7.1"
numpy = "^1.19.2"
torchvision = "^0.8.2"
opencv-python = "^4.4.0"
pytorch-datastream = "^0.4.0"
pytorch-lantern = "^0.11.0"

[tool.poetry.dev-dependencies]
guildai = "0.7.1"
data-kale = "^0.1.3"
pylint = "^2.6.0"
flake8 = "^3.8.4"
black = "^20.8b1"
pytest = "^6.1.2"
notebook = "5.7.8"
jupyter_contrib_nbextensions = "0.5.1"
ipykernel = "^5.4.3"
ipywidgets = "^7.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
